I just made it worse
I have no idea: adding listener on instance.element() doesn't work because it's not focused on (focused on <a> instead), forcing focus doesn't work, listener requires a function without parameters(?) so it's just adds clutter with a slim benefit of removing the listener. I will never use this again in my life.